Decrease font sizes, use CM as the serif font, unify font-switching
To further increase the amount of code that fits in 25 pages, the font
sizes have once again been decreased.

In order for the yet-to-be inter-code texts to be consistent with inline
math, the serif font family has been changed to CMU Serif (TTF version
of the Computer Modern serif fonts).

Two new macros have been defined, \seriffam and \monofam. These macros
simplify the process of switching between the serif and mono fonts as
they redefine \rm, \bf and \it accordingly.