Verbatim: use | as a delimiter, make inline verbatim more usable
If there were no delimiter, "\verb something\egroup" would print an
excess whitespace. This is a hotfix, there should be a more obvious way
of doing it.

Also, a macro \simpleverbatim has been defined, which is meant exactly
for the purposes of inline verbatim, since it doesn't mess with
inter-line spacing.